move api client to its own folder
