+++ ui/app/index.jsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import App from './App.jsx';
 import registerServiceWorker from './registerServiceWorker';
 
-import APIClient from './APIClient'
+import APIClient from '../../client'
 
 const api = new APIClient()
